requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-telegram-bot==21.6
python-dotenv==1.0.1
python-json-logger==2.0.7
cryptography==43.0.1
web3==6.20.1
hexbytes==0.3.1
httpx==0.27.2
